agent: manage state version
Currently, we set the stateVersion of NixOS to a fixed value which
can cause unwanted changes on upgrades. The stateVersion is meant
to be managed separately from the installed NixOS version and only
to be changed when we are sure that it makes sense to update it.
Let the agent write the current system version to a file and read from
it in our NixOS code.
Missing file is still ok, default is the current version.
The file will be required later, failing the system build.
NixOS 22.11 will make the state version a mandatory setting.

The state version could be managed by automated deployments so
the file is placed at /etc/local/nixos/state_version.

 #PL-130893