k3s frontend: adapt for coredns 1.9
CoreDNS removed support for wildcard queries (like
*.myapp.mynamespace.svc.cluster.local) so we cannot get the pod IPs
anymore from non-headless services to load-balance in haproxy.

There are 3 different port options now which are mutually exclusive and
have to match the used service type. This also means that we cannot have
the old fallback concept anymore where we talked to pods directly or
used the cluster IP as backup. The preferred way is now to have an
headless service which still allows us to get the Pod IPs but with a
normal DNS query, which corresponds to the `podPort` option.

Forwarding to cluster IPs (`servicePort`) and to the k3s load balancer
(`lbServicePort`) is also possible.

Documentation also got a significant update, mainly to describe the
networking scenarios more clearly.