fc-ceph-clean-deleted-vms: continue at errors
When encountering errors at deleting rbd images or snapshots, or
creating new rbd pools, the service does not immediately fail but
continues processing the remaining tasks. This allows unaffected tasks
to still be executed within their deadline and not being blocked.
The errors encountered are still printed to the log and the overall exit
code of the service remains a failing one, to make the issues appear in
monitoring.