[hotfix] fc-ceph: no automatic WAL backups at de/activate
The automatic backup and restore at `fc-ceph osd deactivate` and
`fc-ceph osd activate` time for Bluestore OSDs with external WALs was
not resilient against improper OSD deactivation due to e.g. crashes.
This could lead to restoring invalid/ stale/ bagous WAL backups, making
the OSD unusable and requiring it to be completely rebuilt.

In the future, WAL backup and restore will be part of a dedicated OS
migration command.