fc-postgresql upgrade: use --clone flag if possible
Speeds up upgrades of large database clusters by using the reflink
feature of XFS (copy-on-write) which is activated by the `--clone` flag.

Available when target version is > 11, file system supports reflinks
and target data dir is on the same file system. The command
automatically checks if --clone is usable and reports it.

 #PL-131056