Replace upstream docker test with our own
The upstream test doesn't catch dependency changes from over
overlay because it's using a pkgs argument in the test machine
config that doesn't have our overlay. We noticed that problem
after patching sqlite.

The new test also tries a bit more, like talking to a real
application in the container.