statshost-relay: fcio nginx service, log rotation
Enable the flyingcircus nginx service instead of the upstream one.

On most machines running the statshost-relay (rg-relay), webgateway
is also running which enables the flyingcircus nginx service. This
service sets up log rotation for the whole /var/log/nginx dir.

If webgateway is not enabled, this role currently behaves differently,
using the upstream nginx module which means that there is no log rotation
for the relay access/error log. This change unifies the behaviour.

 #PL-131140