nginx: fix allow/deny rules for ipv6 + keep-alive
When keep-alive is on, newer nginx versions seem to recycle data structures our old
anon patch changes. This caused failures because allow/deny
rules were evaluated against the anonymized IPv6 address instead of the original one.
The first request always worked as intended, but subsequent requests did not.

The new patch copies the IPv6 address instead of changing it in-place.