garbagecollect: Keep outputs for build-time deps
Currently, we throw away nixpkgs, nixos-mailserver and more on every
GC run which are then re-downloaded by the next fc-manage run.

We have seen strange issues like partially deleted nixpkgs trees when the GC
and nix-build run concurrently. The GC deletes those temporary store paths
(they are not referenced anywhere after build) but nix-build needs them at
the same time. Note that there are other paths which are not kept even when
`keep-outputs` is switched on, still, so this is not a complete fix.

Keeping them also saves net/disk IO at the expense of some storage space.

PL-131232