fc-agent: integrate garbage collection code and use locking
We have seen weird Nix store issues with concurrent gc and fc-manage runs, so
the new code uses the global agent lock to prevent this. There still can be
concurrent nix-build runs we don't control but at least we won't cause
trouble with our own management timers.

fc-collect-garbage.py was a standalone script which makes more sense as part
of the agent code. It can now be used interactively as `fc-collect-garbage`.
The move also brings proper logging, command line argument handling/help via
Typer and automated testing.

Rename files containing "garbagecollect" because the word order
collect-garbage is already used in more places and it's less confusing to use
the same order everywhere.

The userscan exclude file is now linked to /etc/userscan/exclude which is also
used in the userscan manpage to make it usable for interactive calls of
fc-userscan.

PL-131232