We revamped backy's poetry2nix integration.
Due to a bug in poetry2nix some package changes (like setuptools-scm)
became incompatible. To get those fixed we're now pulling in Python
3.10 and the surrounding ecosystem (poetry, poetry2nix) via our upstream
22.11 tracking repository.

We also stumbled over the sqlite build: the patch was already applied -
I'm not sure why we're running into this in this change, but I had to
remove it to get it building again.