slurm: consider external dependency machines for maintenance
Machines use the role slurm-external-dependency to signal that
they have to be in service in order for the cluster to run jobs,
for example databases. These machines are checked by the controller
for their service state before nodes are enabled. If any of these
machines are not in service, no nodes will be set to ready.