libarchive: disable bsdcpiotests due to being broken on hydra only
While building libarchive on a testvm suceeded, the same build failed on
our hydra machines due to the bsdcpio test not passing. The reasons are
not entirely clear, the symptom was that the `bsdcpio` executable used
by the test was still a libtool wrapper script that prefixed the program
name and output with `lt-`, causing the test output to differ.

I checked that the bit-wise build result content is the same of the
package built with succeeding tests on a testvm and with disabled tests
on hydra. It shall be safe to disable the tests, as at least hydra
appears not to alter the actual package output results.