Restore build compatibility with Nix 2.3/ NixOS 21.11
To ensure upgradability from NixOS 21.11 to 22.05, the `builtins.ceil`
had to be removed as it is only introduced in Nix 2.4, while NixOS 21.11
still evaluates the config with Nix 2.3 at upgrade time.

Semantically, and with the currently chosen rotation times, this does
not change anything.

Fixup for 9992ef28ef3af9d53afeda153804604a3a991824