rgw: supply mime types file
By default, radosgw is looking for mime types specifications under
`/etc/mime.types`, which does not exist.
This fix is not really important, as it is only used by the SWIFT-API
which we do not use at all, but creates a warning at startup that can be
a misleading red herring.