postgresql: revert changes that forced an unneeded restart
Together with the last commit which reverted the last try to
avoid the service restart, this actually fixes the problem.

The README change wasn't the problem because it's just a symlink
which doesn't change the hash for the config directory when the content of
the README changes, so the "fix" in 369235689af was ineffective.

Before all the postgresql changes for the current release, we always added the
versioned config dir to the postgresql.conf, even if there was only the
README.md symlink in it. Since 1ae8c4e7, only dirs with *.conf files were
included, resulting in a postgresql service restart for all machines which
didn't have *.conf there.

Now, we again include the dir regardless of the content.

PL-131380