Add nixos-version file, adapt scripts
nixos-version contains the (major) version of NixOS which we also
use for our platform.

This is helpful for release scripts and similar things to avoid
repeating ourselves and forgetting to change the version everywhere
on a platform upgrade.

fc-release.sh now uses this file.

fc-get-current-channel-url.sh uses it, too.

`production`, `dev` and `staging` can be specified as the only argument to get
the final (redirects resolved) URLs for the branches.

Note that the URLs returned for `production` might already be a newer version
than the one communicated in the changelog as persistent release URL.