platform: reset IPv6 address generation sysctl upon link configuration.
On first boot hosts come up with dhcpcd enabled for dynamic network
configuration. dhcpcd is stopped when switching to the new host's configuration
for the first time, however sometimes dhcpcd removes IPv6 link-local addresses
when it exits and the kernel doesn't replace them.

Setting this sysctl before the interface is configured as part of regular
system activation for the first time should restore the default kernel
behaviour, even if dhcpcd has removed the link-local addresses.