Fix flakiness of mongodb test
The test often fails the auto-restart subtest when IO load is high
on the build machine. The main problem was the `kill -11` which sends
a SIGSEGV. This signal triggers a core dump which may take a long time
when IO-constrained and the 5 second sleep was often not enough.

Now, the subtest uses KILL which is faster and waits until systemd
reports a restart via the NRestarts property. This should always work
without the need for `sleep`.