platform, release: rename network interfaces during stage 1 boot
Rename interfaces in the initramfs, which will prevent race conditions
in later boot stages between sysctls and other udev rules which rely
on interface names.

This requires drivers for network devices to be present in the
initramfs at boot time, so they can be discovered during stage 1. The
udev rules for interface renaming are also repeated in the stage 2
udev configuration, so that if a driver is missing the interface will
still be eventually be renamed and configured correctly.

PL-131741