Implement mechanism to avoid simultaneous maintenance of machines
Add a new NixOS option
`flyingcircus.agent.maintenanceConstraints.machinesInService`
which is used to specify other machines that are not allowed to be in
maintenance mode at the same time when a machine is trying to enter
maintenance.

This is implemented by`fc-maintenance constraints` which takes multiple
`--in-service` arguments to ask the directory API about the "in service"
status of the given machines, returning EXIT_POSTPONE any condition is
not met. The command is configured by the NixOS option to run as
maintenance enter command.

This adds basic user documentation about our maintenance system
which also shows how to use the new option.

PL-131290