rabbitmq 3.6.5 role for upgrades from NixOS 20.09
This is a new type of role: it only exists for upgrading existing VMs
directly from NixOS 20.09. The rabbitmq service from the old version is
"frozen" by generating local config that just takes the existing unit
file and its dependencies and adds it as a systemd package. This way,
the service doesn't even restart when the system is upgraded to 23.11 as
nothing changed for the service.

The added platform code is mostly just lifted from the 20.09 rabbitmq36_5
role and the corresponding service.

PL-131333