tests: adapt guest vlan handling for upstream test library changes
As a side-effect of the introduction of the upstream
virtualisation.interfaces option for explicitly specifying the names
of interfaces in guest VMs, interfaces configured using the existing
virtualisation.vlans option now include udev rules which force the
name of the interface to ethX (for some vlan X). Due to ordering these
rules override the rules set in the platform, which means that network
interfaces are not named as expected within test VMs.

As there is no straightforward way to disable these extra udev rules,
refactor the tests to use the new virtualisation.interfaces option to
specify the required network interfaces with their desired names.