k3s: allow kubelet to read DaemonSets
The kubelet has permissions to manage pods and to amend node status,
which is nearly enough for performing a node drain operation with
kubectl. As draining handles DaemonSet pods specially, add the
required permissions to read DaemonSets to the kubelet's access
control role, so its credentials can be reused by the maintenance
automation for evicting workloads before entering maintenance.

PL-131525