agent: remember release metadata from ENC and use it
Release metadata is now read from enc.json and written to releases.json,
with the version label as key. The data can then be accessed via a
read-only NixOS option flyingcircus.platform.release which is used to
add the proper changelog URL and release name to the motd file. This
file can now also be accessed at /etc/motd.

This also adds the changelog URL and release name to the summary for
an UpdateActivity, which is used as maintenance request comment.

PL-131736