fc-ceph/osds: extract OSDs into separate units to avoid shared fate
We had hosts run out of memory and the OSDs were all subsumed in the
same cgroup which caused them to be killed together because systemd
values consistency in a unit's state ...

There might be interesting cases here if NixOS decides to
activate/deactivate anything, but hopefully the current setup will cause
the "group unit" to catch this. I can imagine cases where this might
not be the case, but I can't show that they actually exist currently.

Re PL-131646
co-authored by os@flyingcircus.io