slurm: wait for all nodes being ready
Implement ready_many similar to drain_many which waits until the
desired state is reached, with a configurable timeout. If nodes
failed to become ready, they are reported in the timeout exception.
Usually we expect the nodes to become ready almost immediately so
there shouldn't be a noticeable waiting period.

ready_many is now used by the `all-nodes ready` subcommand.

PL-131739