agent: do not acquire lock when getting metrics
We also want metrics during a maintenance run.
It's generally safe to call ReqManager methods as long as
they don't modify request state.

__enter__ now only acquires the lock. Reading the config is
moved to __init__ so that methods can read config without
locking.

This also includes a fix to include pytest in the Python dev
environment again as the pytestCheckHook doesn't propagate pytest
itself.

There's a possibility that loading requests will lead to an error
which means that a "read-only" method might change state because
requests that failed to load are archived immediately. This should
be fixed later as it can be annoying for testing/debugging.

PL-131185