fc.qemu: improve orphaned log rotation, increase / size for physical
We noticed that the log rotation doesn't work cleanly for logs in
/var/log/vm because VMs are migrating back and forth and we end up
with traces of logs from every VM on every KVM host. In addition
logrotate isn't aware of VMs that have left and isn't able to see
whether old files are still being used.

This script adds another layer to detect orphans and clean them up
safely.

In addition, we increase the default size of the root partition for
physical hosts from 40 to 80GiB.

Fixes PL-131715