agent: add missing UpdateActivity attributes on deserialization
We added some attributes to UpdateActivity which may cause exceptions
when old serialized instances are loaded by newer agent code and the
new code expects them to be present.

Often, this isn't triggered as UpdateActivities are archived after the
update and not touched again but as these activities can do reboots or
may fail (typically on the system switch), they might hang around for a
longer time and be loaded again by updated agent code.

This also acts an on-disk "schema migration" of old activities as
requests are loaded and saved back by the reqmanager if something
changes.

PL-131774