agent: treat UpdateActivity switch problems as temporary failure
As we treat every error when running an update activity as permanent,
we often get maintenance error notifications for updates when the first
switch fails. The second try usually succeeds when fc-manage runs the
next time but the error mail is already sent.

Switch errors are now treated as temporary failures which are retried on
the next maintenance run.

This change also addresses unwanted behaviour when an update needs a
reboot because of a changed kernel and then fails in the switch phase.

Activity failure prevents the system from rebooting immediately, so
there's difference between running and wanted kernel.

Before, the next `fc-maintenance request` run would then schedule a
reboot which is now unneccessary as the update activity is still present.
The activity will trigger the reboot when the retry succeeds.

PL-131769

(cherry picked from commit 7afea8e53a27aa205df1e1b39681d683bfeed618)