nginx and acme: fix failing acme validation at first run
This workaround ensures that newly added nginx vhosts are picked up
before the acme validation service runs and requires them.
After the validation, nginx still needs to be reloaded (again) to pick
up the new certificates.

We might be able to find a better solution upstream, but as we have
forked the nginx module anyways this provides a quicker fix.