ceph: refactor versioned packaging
The kernel upgrade broke the KVM test and the current way of
correctly accessing packages that depend on Ceph with our multi-version
approach (to assist migration scenarios when we have multiple
Ceph versions around) was extremely confusing.

Try a new factoring where there is a single set of all packages
that use a specific version of Ceph (mkPkgs) but do not hard-code
any details about how those versioned packages are tried together.

We expect to wire the individually versioned packages up "correctly"
in our overlay and expose them with a standardized naming scheme so
that the function can pull them together.