lamp package test: use the correct nixpkgs, not the bootstrap nixpkgs
Hydra should build/test the lamp packages for the nixpkgs version pinned
by fc-nixos. Before, the test pulled in pkgs from the bootstrap nixpkgs
which is often different as it is updated separately in Hydra jobset
settings.

This change is mostly for completeness, the lamp VM tests already did the right
thing and make sure that the packages are built on Hydra.

PL-131831