ceph-nautilus: Raise log level for informational messages
For getting an RGW access log, we need to set "debug rgw" to at least a
level of 1. Unfortunately, this also gives us a large number of
informational or notice logs in OSD logs.
For all RGW-class log messages not clearly related to errors or
cancellations, I increased the required log level to 2. This helps
differentiationg between an error-like and info-like log.

Note: some of the messages now ignored had the format
`read_index_entry(): existing entry:`

I hereby decide that (already) existing bucket entries are not relevant
for error-like logs.