agent: handle legacy reboot activities properly
Old reboot activities rebooted during activity execution which we don't
want to do anymore with the new ReqManager design. It's possible that
they have been created before switching to new agent code which tries to
execute them, leading to an unexpected reboot. Instead, we now use the
new implementation which is basically a no-op and defers rebooting to
the ReqManager.

PL-131849