agent: implement resuming interrupted requests
Before, requests that had been persisted in state "running" were just
ignored by the ReqManager and had to be cleaned up manually.

Now, the default behaviour defined in Activity.resume() is to put the
activity into an error state (special value EXT_INTERRUPTED = 80) to
avoid re-running non-idempotent activity that could be unsafe to run
again.

Idempotent acivities (UpdateActivity, VMChangeActivity) are just
run again.

The legacy reboot activity is a special case: it was usually interrupted
by the planned reboot triggered in activity.run() so we just mark it
as successful to clean it up.

PL-131849