agent: command logging for update activity runs, ignore main log file permission errors
Output of nix commands in UpdateActivity.run() is now logged. A typical
problem are failing services on system activation which we want to be
logged to a separate file like for `fc-manage switch`.

Main log file permissions problems are now ignored.
No need to abort when the main log file cannot be used.
We still have journal logging and console output.

PL-131813