mysql: suppress deprecation warning for mysql_native_password
This warning is new in 8.0.34 and is logged for every connection attempt
which can be a lot of log spam for applications that create a new
connection on every request. It's already annoying for our telemetry
accounts and we cannot change the password type as the sensu plugin only
supports the deprecated one.

PL-131639