Use Linux 5.15
We experience a known mysterious kernel bug on 6.1, in the memory
management (mm) subsystem after large folio support was introduced in
5.18. We cannot reproduce the bug and there's no idea how to fix it. We
have only seen it on VMs with postgresql and percona/mysql roles. 5.15
is a bit slower without the large folio support but the bug already
caused file system corruption and we don't want to risk that again.

PL-131574