release: fix excluding nested packages, refactor modified pkg jobs
Jobs for "modified" packages from our overlay (it's all packages from
the overlay actually, not only the modified ones) are now generated from
the same code path as the jobs for important packages.

To be able to use that code, packages from the overlay are turned into a
list of "dotted names" like `adns` or `fc.blockdev`. This allows us to
exclude specific packages from nested sets (only `fc` at the moment)
using the dotted name. The previous code only supported excluding
things on the top level of the overlay.

The exclusion list is now used for both important and overlay packages.

PL-131581