agent: do not repeat existing comment content when merging
Now, when merging requests the comment of the new request
is only concatenated to the old one when it's not contained in the
old commit to avoid repeating content. We saw this with RebootActivity
which is created on every fc-agent run when a reboot for the kernel is
needed. When the kernel version changed again, the new comment was
concatenated over and over again.