update-nixpkgs, versions: sha256 -> hash, fix rsync problem
Also make sure that the globally-installed nix-prefetch-github on
hydra01 is used as users might have different versions of the script.

The rsync cmd for package-versions now makes sure that the target
directory is specified correctly. Depending on how the command was
called, it was possible that the content was accidentally synced to the
parent dir.