varnish: fix the multihost configuration
the multihost configuration was broken because although the correct
configuration was applied initially, during a reload of the varnish
service a blank configuration would be loaded because the reload
script was not written with the legacy case in mind.

in order to fix this, and to simplify the varnish service logic,
the legacy config from the "/etc/local/varnish" directory is now
rendered as a virtualhost that accepts any incoming request

this, however, means that a case where there is a legacy config file
as well as nix-defined varnish vhost configuration is harder to
detect, which is why the corresponding assertion has been removed.