platform: relax AS-path constraints for multi-path routing
The default behaviour of EBGP is such that two routes with different
AS-paths are not considered equal, as on the public internet routes
with different AS-paths will traverse a different set of
networks (i.e. administrative domains), even if they're the same
length. Only routes which have the same AS-path are considered
eligible for merging into a single multipath route.

In the datacentre these conditions do not apply: in spite of using
EBGP for exchanging routing information, all nodes are part of the
same adminstrative domain. Hence only the lengths of the paths are
relevant for calculating multi-path routes, and not the AS's within
that path.