upgrade docs: remove info about production status
It's too easy to forget updating it and even finding a good place in time to
change it isn't trival due to our release process. We communicate in the
changelog when platform versions should be used for non-prod/prod, maybe
als adding information about this to the customer UI would be helpful.