roles/mailserver: require an actual url as imprintUrl
PL-132155

Even though the option was called `imprintUrl` the protocol scheme was
prepended unconditionally. This means when defining an `imprintUrl` like

    https://example.org/impressum

a redirect to `https://https//example.org/impressum` will be issued
which is pretty unintuitive.

Since the option claims to accept a URL I decided to change its
semantics like this:

* A URL is expected now (as long as it's accepted by the `return` clause
  of nginx).

* For backwards-compatibility, a "URL" w/o a scheme (i.e. `example.com`)
  is also accepted. This behavior is deprecated however and gives the
  following warning at evaluation time:

      trace: warning: Specifying flyingcircus.roles.mailserver.imprintUrl without a protocol scheme is deprecated. Add https:// to the declaration (i.e. "example.com" -> "https://example.com") to get rid of this warning.

  We could drop the old behavior after e.g. the first platform release
  with NixOS 24.11. When doing that, the `apply` expression should be
  dropped and the validation should be done by using
  `nullOr (strMatching "")` instead.

A test-case to the integration test of the mailserver role was added
which ensures that the redirect declared by `imprintUrl` works fine.