sensu: fix check_ping
This worked until upstream changed the hard-coded paths for ping from a
/run/wrappers path (doesn't exist on 23.11 anymore) to just 'ping',
relying on PATH. For check_ping, this doesn't work as PATH is
ignored. Instead, we can hard-code the full path to ping during build
time.

In addition to that, this change includes a comment about additional checks added by
us which we don't use in platform code. We can probably remove them in
24.05.

postInstall is also overridden in a safe way now, appending to it if it
already exists upstream.

PL-132185