nixos/dev-vm: run varnish as dualstack to be consistent with production VMs
PL-132197

In some deployments we use IPv4 addresses for the
nginx->varnish->haproxy chain. This breaks when talking to varnish
in a dev-host setup however as it listens to IPv6 only in a
dev-host environment which is inconsistent with the production VMs.

`::` isn't enough for that, Varnish forces `::` to be IPv6-only[1].

[1] https://github.com/varnishcache/varnish-cache/blob/223847ea09a0a5440510fe8c383bdac56890dff5/lib/libvarnish/vtcp.c#L448-L460