fc.ceph: refactor LVM abstractions for modular base block device
Refactor the LVM volume management abstractions to work on top of different block devices instead of plain disks as a base device.
For this purpose, `GenericBlockDevice` is introduced as an abstraction. The concrete implementation `PartitionedDisk` retains the existing behaviour and makes the ceph NixOS tests pass.
MdraidDevice is just a stub so far.