roles/gitlab: enable CSP by default
PL-132241

See https://docs.gitlab.com/omnibus/settings/configuration.html#set-a-content-security-policy

By enabling this, the Rails application will generate a proper CSP
header in accordance to what assets are delivered. From now on, this is
the default for each GitLab.

Confirmed that this behaves properly by testing it on our test instance
on fcstag60. Also wrote a small integration test to make sure that such
a header is always delivered with a non-empty value (currently, the
header already exists according to curl, but is empty).