agent: check free disk space for updates
request_update now guesses if the system to be prepared will fit on
the Nix store partition. It uses the size of the current system plus
some extra space as worst-case estimation. This should leave at least
some free space after preparing the new system.

Often nixpkgs updates change core dependencies, so adding the size of the
current system when preparing a new one is not uncommon.

We might miss some smaller updates with this estimation and it could be
relevant for security. We have to make sure to watch the fc-agent check
which will become red when there's not enough space.

Additions:

- Successful system builds now show the closure size of the new system.
- `fc-manage check` now:
  - displays current system closure size as info.
  - warns when less than twice the system size plus some extra space is free.
  - errors when less than the system size plus some extra space is free.

PL-131589