ceph osds: prevent them from being stopped by fc-blockdev restarts
This fixes the issue of all OSD service units being shut down and staying down when fc-blockdev.service was stopped (and started) again by a system switch.
The main issue was that stops of fc-blockdev propagated to stop the individual OSD units, but nothing ensured that they come up again. Only fc-ceph-osds-all.service was pulled in by the multi-user.target and stayed active all the time. By reducing the dependency type towards fc-blockdev, we prevent the osds-all unit from being stopped in the first place.

One side-effect of this is though that OSDs will start even if
fc-blockdev.service fails.

PL-132304 PL-132303