ceph osds: ensure they come up again after fc-blockdev restarts
This fixes the issue of all OSD service units being shut down and staying down when fc-blockdev.service was stopped (and started) again by a system switch.
The main issue was that stops of fc-blockdev propagated to stop the individual OSD units, but nothing ensured that they come up again. Only fc-ceph-osds-all.service was pulled in by the multi-user.target and stayed active all the time. By adding the fc-blockdev requirement to the osds-all.service, it is ensured that the stopping of OSDs is discovered and they are being activated again after restart.

Stoppping osd services at a system switch is a notable change we've tried to avoid for now. But doing so makes sense, as it allows automatically applying changed blockdevice settings.
Doing OSD restarts for further changes like config file or executable changes is out of scope for now and can be revisited at the next ceph update.

Another important consideration to be made: OSDs explicitly stopped won't always stay stopped, as they might be started again when fc-blockdev changes. But this could already happen when a machine had scheduled a reboot, as after boot all OSDs are started.

PL-132304 PL-132303