network: refactor interface configuration units
Upstream NixOS has a convention of naming systemd units which manage
virtual network interfaces "IFACE-netdev.service". Here we adopt this
convention, and rename "network-link-properties-vxFOO-virt.service" to
"vxFOO-netdev.service" for consistency.

A variety of network configuration related units from
upstream (notably those performing address configuration) have
"stopIfChanged = false" set. This means these units will be restarted
in-place to minimise the risk of network interruption during
switch-to-configuration (instead of being stopped before the new
system is activated and restarted afterwards). This flag was not set
on platform-provided network units, which could have led to problems
where units which expected to be restarted (or reloaded) in-place were
instead stopped due to dependency units being stopped.

Additionally, the code for configuring network interfaces when VXLAN
is enabled has been reorganised, in order to hopefully be more
digestible to the passing reader.

PL-132305