ceph: workaround for 8ab330b2d48dc88a5b9698d191b75e0086fdcc0d
Unfortunately the changes in 8ab330b2d48dc88a5b9698d191b75e0086fdcc0d
are not sufficient for preventing the shutdown of OSDs, as the decision
at switch-to-configuration time on which dependent units to stop when
fc-blockdev is stopped relies on the unit state _before_ the switch. As
long as the `requires` relation between OSDs and fc-blockdev is still
there in that state, OSDs will still be stopped and do not come up again
automatically.

This utilises a trick of using a `systemctl restart` instead of a stop
and start in switch-to-configuration, as the restart is already done
based on the new unit file and its dependency structures.
Also, changing this modifies the unit file itself and already triggers a
restart of fc-blockdev.service by itself, making sure the changes are
persisted.

This _might_ be removed in the future again, as its main purpose is
making sure the transition of fc-ceph-osd@.service dependencies happens
cleanly without pulling down the osds.