network: restart zebra when loopback addresses change
zebra uses the address configured on the ul-loopback interface as the
source address for underlay routes it inserts into the kernel. If this
source address is changed or removed, then zebra will remove these
routes from the kernel and fail to reinsert them, and will not recover
from this condition without manual intervention.

Ensure that zebra is restarted when the loopback interface's assigned
addresses are changed, to force this condition to be cleared.

PL-132305