percona 8.1: package, role, test, xtrabackup
The xtrabackup package expression is a separate file, as we had an issue
with the percona-xtrabackup 2.4 source being pulled
from binary cache for some reason.
This should be resolved before bringing this upstream to nixpkgs.

We've already upstreamed percona 8.0 to nixpkgs
and pull it in from there, thus the orphaned
percona_8_0 expression file can be removed.

Co-authored by: Oliver Schmidt <os@flyingcircus.io>